requirements.txt
fastapi==0.78.0
azure-cosmos==4.2.0